New deal, you get your free flips when you get your new grid, no funny accumulation. Have all your fun when you play your new grid and then go do something else for 24 hours.
Also, you get 5 free flips per day instead of 3. w00t!


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@188 a9e240f3-4568-de11-90a4-003048321218

+++ src/java/com/threerings/everything/data/PlayerDetails.java
@@ -35,5 +35,5 @@ public class PlayerDetails
     public int coins;
 
     /** This player's current free flips status. */
-    public float freeFlips;
+    public int freeFlips;
 }
